Do not modify the input buffer, but rather use the output buffer
After a code review from Justo it is clear that the solution implemented
was modifying the input buffer, which was then written untouched to the
output one. However a better solution is not to touch the input buffer
and store the modified buffer (with the sigmaToWeight function) in the
output MS directly. This mimics what is being done already for the
weight to sigma transformation.
Fixes partially CAS-11139