Reenable tests with time average, checking weight and sigma
First of all a test was added to compare time average data with averaging
interval equal to data interval: 30 seconds. In this case it is
expected that the weight and sigma are the same as with no time
averaging

Then checks have been enabled in the case of real average, with 60 seconds
in the case of column='DATA'
The array values that were there commented out have been changed.
The new ones seem very consistent
with the expectation: The interval of the data is 30s and the
time average interval is 60s, therefore one expects the new
sigmas to be the same as the old one divided by sqrt(2). Take
into account that for the test data the sigmas are the same in the
two timestamps. For instance, for the first timestamp sigma
is [4, 3, 2, 1] (row 0 in MS) as well as for the second (row 15 in MS).
After averaging and dividing by sqrt(2) one ends up with
[2.82, 2.12, 1.41, 0.70]

Finally, and simila to previous check, further checks for
weights and sigma columns are checked in the case column='CORRECTED'.
In this case the weight column is the relevant one,
and sigma is computed from it. The values also have to be changed accordingly.

Fixes partially CAS-11139