This commit is to enable antenna group finding given a sigma threshold for binning antenna pointing deviations from the median. The is done in the following fashion steps.
1. PointingOffsets.cc has an additional method PO::fetchAntOffsetToPix
which will return the antenna offset from the vb direction.
2. This allows us to determine per vb the individual antenna offsets
which is done by VB2CFBMap::findAntennaGroups
3. Once the groups are identified it returns the information to the
control method VB2CFMap::makeVBRow2BLGMap which returns a list of baseline
groups per row of the vb given an antenna. The groups are computed using
the formula NBLG = NG + (NG C 2).

This commit does not include the plugin to activate this code. That will
follow shortly