Full support for user-defined sorting in subchunks
With this change the VI2 supports a fully user-defined
way to iterate within the chunks (i.e., the subchunk iteration).
If the user does not define (with the new constructor) the
sorting for subchunks, then the current behaviour is kept: provide
all rows with a given timestamps in each subchunk.

Note that commit fba861d11c2 also contains part of this work, which was
commited inadvertently.

Fixes partially CAS-12879