Define validity scope for metadata.
Some metadata can vary at different rates. For instance,
the DDID can be unique within a chunk, a subchunk or in
the worst case be valid only for a row. This adds initial support
in the VI/VB2 framework to determine how long a given metadata
can be considered valid without having to recompute (which might be
pontetially expensive).
Fixes partially CAS-12895