Add new API functions to retrieve vector of cubes with different shape
The new API functions can be used while iterating the MS to retrieve
all the subchunk in the case that potentially contains rows
with array columns that have different array shapes in each row.
A vector of arrays (cubes/matrices) is in this case returned.
The new API could also be used in the "normal" case with just a single
shape across the whole subchunk. The returned vector would have a size
of 1.
Fixes partially CAS-12895