Implement the new API functions for vectors of arrays.
This commit adds support in VisibilityIteratorImpl2 to the new VI2 API
functions that allow retrieving vector of cubes with different shape
each.
Since the list of channel selectors has already been created in previous
commits the only remaining task is to loop for each different
shape and call the column accesors with the proper slicer.
Fixes partially CAS-12895