Use rownr_t in places where the number of rows in the VisBuffer are expected
The nubmer of rows in a VisBuffer2 can potentially be as large
as the whole table, depending on how the iteration grouping is performed.
Therefore everything related to the number of rows in the VisBuffer2
should be able to accomodate as many rows as a table can, i.e., the type
should be rownr_t.
Note that the VisBuffer2 access the values via an array, for which the
maximum size is determined by type size_t, so there is a potential
mismatch here between size_t and rownr_t. For further discussion
see https://github.com/casacore/casacore/issues/1037.

Fixes partially CAS-13137