Fix pb math in task_simalma (CAS-13369)
The .image file produced by clean in 5.4 had primary beam correction applied. Since CAS-12502 and the change to using imtclean instead of imclean in simanalyze (called by simalma) the .image file is now sky-flat and does not need an extra scaling by the primary beam before it is feathered with the SD image.

Ensure the task_simalma code follows the algorithm:
* create SD image
* create IF flat-noise image
* regrid SD to IF
* output = feather(SD*IF.pb, IF(flat-noise) )/ IF.pb